# specs/003-dashboard-fix/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dashboard Application API
  description: API for the dashboard application with task management, projects, tags, and notifications
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourapp.com
    description: Production server
paths:
...
  /api/v1/tasks:
    get:
      summary: Get all user tasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
  /api/v1/tasks/{task_id}:
    get:
      summary: Get a specific task
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    put:
      summary: Update a task
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted successfully
        '404':
          description: Task not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [pending, in-progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date-time
        projectId:
          type: string
        userId:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - userId
    TaskCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [pending, in-progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date-time
        projectId:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - status
        - priority
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [pending, in-progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date-time
        projectId:
          type: string
        tags:
          type: array
          items:
            type: string
...